+++ CHANGELOG.rst
@@ -3,10 +3,11 @@ dev
 
 Release date: ``2017-??-??``
 
+-  `NXPY-11 <https://jira.nuxeo.com/browse/NXPY-11>`__: Add usage examples
 -  `NXPY-14 <https://jira.nuxeo.com/browse/NXPY-14>`__: Quote Repository URLs
 -  `NXPY-22 <https://jira.nuxeo.com/browse/NXPY-22>`__: Sanitize relative URLs
 -  `NXPY-25 <https://jira.nuxeo.com/browse/NXPY-25>`__: Allow strings as properties
--  `NXPY-27 <https://jira.nuxeo.com/browse/NXPY-27>`__: Use of setup.cfg
+-  `NXPY-26 <https://jira.nuxeo.com/browse/NXPY-26>`__: Use of setup.cfg
 -  `NXPY-29 <https://jira.nuxeo.com/browse/NXPY-29>`__: Fix an encoding error in ``Nuxeo._log_details()``
 -  A lot of code clean-up and improvement
 
@@ -21,6 +22,7 @@ Technical changes
    nuxeo/blob.py::\ ``get_upload_buffer()``
 -  Moved ``FileBlob._read_data()`` to nuxeo/blob.py::\ ``_read_data()``
 -  Added ``Nuxeo.drive_config()``
+-  Changed ``Nuxeo._check_params()`` to ``Nuxeo.check_params()``
 -  Changed ``Nuxeo._rest_url`` to ``Nuxeo.rest_url``
 -  Changed ``NuxeoObject._service`` to ``NuxeoObject.service``
 -  Added ``NuxeoService.exists()``